+++ src/App/components/Tabs/index.test.js
@@ -3,16 +3,7 @@ import { shallow } from "enzyme";
 
 import Tabs from "./index";
 
-const items = [
-    {
-        name: "test",
-        active: false
-    },
-    {
-        name: "test 2",
-        active: true
-    }
-];
+const items = ["test", "test 2"];
 
 // TODO: Write more tests!! [EH]
 
@@ -37,6 +28,5 @@ describe("Component: Tabs - ", () => {
         );
 
         expect(wrapper).toMatchSnapshot();
-        expect(wrapper.contains(<a href="#">test</a>)).toEqual(true);
     });
 });